rename SharedUtils to CanvasDataConstructor
