+++ src/renderer/components/shared/CanvasDataConstructor.js
